Avoid printing char16_t* and wchar_t to gtest Message
gtest's Message class has a special handler for operator<< of wchar_t*
to convert it to UTF-8, but it doesn't have one for a single wchar_t or
for a char16_t* string. It delegates these to std::stringstream, which
as of a libc++ upgrade, deletes its operator<< for char16_t and
wchar_t. See wg21.link/p1423r3.

Bug: http://b/175635923
Test: m MODULES-IN-bionic

+++ tests/wctype_test.cpp
@@ -36,19 +36,19 @@ static void TestIsWideFn(int fn(wint_t),
   UtfLocale l;
   for (const wchar_t* p = trues; *p; ++p) {
     if (!have_dl() && *p > 0x7f) {
-      GTEST_LOG_(INFO) << "skipping unicode test " << *p;
+      GTEST_LOG_(INFO) << "skipping unicode test " << static_cast<int>(*p);
       continue;
     }
-    EXPECT_TRUE(fn(*p)) << *p;
-    EXPECT_TRUE(fn_l(*p, l.l)) << *p;
+    EXPECT_TRUE(fn(*p)) << static_cast<int>(*p);
+    EXPECT_TRUE(fn_l(*p, l.l)) << static_cast<int>(*p);
   }
   for (const wchar_t* p = falses; *p; ++p) {
     if (!have_dl() && *p > 0x7f) {
-      GTEST_LOG_(INFO) << "skipping unicode test " << *p;
+      GTEST_LOG_(INFO) << "skipping unicode test " << static_cast<int>(*p);
       continue;
     }
-    EXPECT_FALSE(fn(*p)) << *p;
-    EXPECT_FALSE(fn_l(*p, l.l)) << *p;
+    EXPECT_FALSE(fn(*p)) << static_cast<int>(*p);
+    EXPECT_FALSE(fn_l(*p, l.l)) << static_cast<int>(*p);
   }
 }
